requirement.txt
asgiref==3.7.2
Django==4.2.1
djangorestframework==3.15.1
packaging==23.1
pyparsing==3.0.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.2
tzdata==2023.3
